Feature/plotting

+++ modules/postprocessing/src/processors/ssao.cpp
@@ -277,7 +277,7 @@ void SSAO::process() {
     auto persp_camera = dynamic_cast<PerspectiveCamera*>(&camera_.get());
     if (persp_camera) {
         projParam_.ortho = 0;
-        projParam_.fov = glm::radians(persp_camera->getFovy());
+        projParam_.fov = static_cast<float>(glm::radians(persp_camera->getFovy()));
     } else {
         projParam_.ortho = 1;
         projParam_.orthoheight = static_cast<float>(height);
